refactor(config): drop experimentalHooks barrel re-exports
Retarget every caller that pulled flag constants, defaults, or
getExperimentalFlag through @/config/experimentalHooks to import
directly from @/config/experimental and @/config/experimentalConfigStore.
Test mocks for getExperimentalFlag move to the new module too. With
nothing left routing through the barrel, remove the re-exports from
experimentalHooks.ts so it only exposes the React hooks.
