+++ src/components/asset-list/RecyclerAssetList2/LegacyWrappedTokenFamilyHeader.tsx
@@ -7,7 +7,8 @@ import CaretImageSource from '@/assets/family-dropdown-arrow.png';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { ImgixImage } from '@/components/images';
 import TokenFamilyHeaderIcon from '@/components/token-family/TokenFamilyHeaderIcon';
-import useExperimentalFlag, { NFTS_ENABLED } from '@/config/experimentalHooks';
+import { NFTS_ENABLED } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { Text } from '@/design-system';
 import useLatestCallback from '@/hooks/useLatestCallback';
 import * as i18n from '@/languages';
