+++ src/components/expanded-state/ContactProfileState.js
@@ -1,7 +1,8 @@
 import React, { useCallback, useMemo, useState } from 'react';
 import { Keyboard } from 'react-native';
 
-import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
+import { PROFILES } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import useENSAvatar from '@/features/ens/hooks/useENSAvatar';
 import { removeFirstEmojiFromString, returnStringFirstEmoji } from '@/helpers/emojiHandler';
 import useContacts from '@/hooks/useContacts';
