+++ src/components/send/SendHeader.tsx
@@ -7,7 +7,8 @@ import isEmpty from 'lodash/isEmpty';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import Divider from '@/components/Divider';
 import Spinner from '@/components/Spinner';
-import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
+import { PROFILES } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import styled from '@/framework/ui/styled-thing';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
 import { opacity } from '@/framework/ui/utils/opacity';
