+++ src/features/charts/candlestick/components/CandlestickChart.tsx
@@ -39,7 +39,8 @@ import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { EasingGradient } from '@/components/easing-gradient/EasingGradient';
 import { DelayedMount } from '@/components/utilities/DelayedMount';
 import { MountWhenFocused } from '@/components/utilities/MountWhenFocused';
-import useExperimentalFlag, { CANDLESTICK_DATA_MONITOR } from '@/config/experimentalHooks';
+import { CANDLESTICK_DATA_MONITOR } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { globalColors, Text, TextIcon, useColorMode, useForegroundColor } from '@/design-system';
 import { getColorForTheme } from '@/design-system/color/useForegroundColor';
 import { useSkiaText, type TextSegment } from '@/design-system/components/SkiaText/useSkiaText';
