+++ src/features/charts/stores/chartsStore.ts
@@ -1,4 +1,5 @@
-import useExperimentalFlag, { CANDLESTICK_CHARTS } from '@/config/experimentalHooks';
+import { CANDLESTICK_CHARTS } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { isHyperliquidToken } from '@/features/charts/utils';
 import { useRemoteConfig } from '@/model/remoteConfig';
 import { createRainbowStore } from '@/state/internal/createRainbowStore';
