+++ src/features/king-of-the-hill/hooks/useShowKingOfTheHill.ts
@@ -1,4 +1,5 @@
-import useExperimentalFlag, { KING_OF_THE_HILL_TAB } from '@/config/experimentalHooks';
+import { KING_OF_THE_HILL_TAB } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { IS_TEST } from '@/env';
 import { useRemoteConfig } from '@/model/remoteConfig';
