+++ src/features/positions/stores/transform/filter.ts
@@ -1,5 +1,5 @@
+import { DEFI_POSITIONS_THRESHOLD_FILTER } from '@/config/experimental';
 import { getExperimentalFlag } from '@/config/experimentalConfigStore';
-import { DEFI_POSITIONS_THRESHOLD_FILTER } from '@/config/experimentalHooks';
 
 import { type RainbowPosition } from '../../types';
 import { DetailType, PositionName, type PortfolioItem, type PositionToken } from '../../types/generated/positions/positions';
