+++ src/hooks/useWalletSectionsData.ts
@@ -3,7 +3,8 @@ import { useEffect, useMemo } from 'react';
 import { analytics } from '@/analytics';
 import { type AssetListType } from '@/components/asset-list/RecyclerAssetList2';
 import { type CellTypes } from '@/components/asset-list/RecyclerAssetList2/core/ViewTypes';
-import { CLAIMABLES, DEFI_POSITIONS, RNBW_REWARDS, useExperimentalConfig } from '@/config/experimentalHooks';
+import { CLAIMABLES, DEFI_POSITIONS, RNBW_REWARDS } from '@/config/experimental';
+import { useExperimentalConfig } from '@/config/experimentalHooks';
 import { IS_TEST } from '@/env';
 import { usePerpsFeatureCard } from '@/features/perps/hooks/usePerpsFeatureCard';
 import { usePerpsPositionsInfo, type PerpsPositionsInfo } from '@/features/perps/stores/derived/usePerpsPositionsInfo';
