+++ src/navigation/Routes.ios.tsx
@@ -8,7 +8,8 @@ import { SwapScreen } from '@/__swaps__/screens/Swap/Swap';
 import { ControlPanel } from '@/components/DappBrowser/control-panel/ControlPanel';
 import { LogSheet } from '@/components/debugging/LogSheet';
 import WalletErrorSheet from '@/components/wallet-error/WalletErrorSheet';
-import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
+import { PROFILES } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import AppIconUnlockSheet from '@/features/app-icon/screens/AppIconUnlockSheet';
 import BackupSheet from '@/features/backup/components/BackupSheet';
 import RegisterENSNavigator from '@/features/ens/navigation/RegisterENSNavigator';
