+++ src/screens/AddWalletSheet.tsx
@@ -12,7 +12,8 @@ import WalletsAndBackup from '@/assets/WalletsAndBackup.png';
 import WatchWalletIcon from '@/assets/watchWallet.png';
 import { AddWalletList } from '@/components/add-wallet/AddWalletList';
 import { type AddWalletItem } from '@/components/add-wallet/AddWalletRow';
-import useExperimentalFlag, { HARDWARE_WALLETS } from '@/config/experimentalHooks';
+import { HARDWARE_WALLETS } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { Box, globalColors, Inset } from '@/design-system';
 import { IS_DEV } from '@/env';
 import { executeFnIfCloudBackupAvailable } from '@/features/backup/backup';
