+++ src/screens/SendConfirmationSheet.tsx
@@ -9,7 +9,8 @@ import { isEmpty } from 'lodash';
 import ShimmerAnimation from '@/components/animations/ShimmerAnimation';
 import RainbowCoinIcon from '@/components/coin-icon/RainbowCoinIcon';
 import Divider from '@/components/Divider';
-import useExperimentalFlag, { PROFILES } from '@/config/experimentalHooks';
+import { PROFILES } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { Box, Heading, Inset, Stack, Text, useBackgroundColor, useColorMode } from '@/design-system';
 import { AssetType } from '@/entities/assetTypes';
 import ENSCircleIcon from '@/features/ens/components/ENSCircleIcon';
