+++ src/screens/SettingsSheet/components/SettingsSection.tsx
@@ -20,7 +20,8 @@ import PrivacyIconDark from '@/assets/settingsPrivacyDark.png';
 import WalletsAndBackupIcon from '@/assets/WalletsAndBackup.png';
 import { AppVersionStamp } from '@/components/AppVersionStamp';
 import { XIcon } from '@/components/icons/svg/XIcon';
-import useExperimentalFlag, { LANGUAGE_SETTINGS, NOTIFICATIONS } from '@/config/experimentalHooks';
+import { LANGUAGE_SETTINGS, NOTIFICATIONS } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { Box } from '@/design-system';
 import { backupsStore } from '@/features/backup/stores/backupsStore';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
