+++ src/screens/change-wallet/ChangeWalletSheet.tsx
@@ -14,7 +14,8 @@ import { EasingGradient } from '@/components/easing-gradient/EasingGradient';
 import { SheetHandleFixedToTop } from '@/components/sheet';
 import { Panel, TapToDismiss } from '@/components/SmoothPager/ListPanel';
 import { FeatureHintTooltip, type TooltipRef } from '@/components/tooltips/FeatureHintTooltip';
-import useExperimentalFlag, { NOTIFICATIONS } from '@/config/experimentalHooks';
+import { NOTIFICATIONS } from '@/config/experimental';
+import useExperimentalFlag from '@/config/experimentalHooks';
 import { Box, globalColors, HitSlop, Inline, Text } from '@/design-system';
 import type { EthereumAddress } from '@/entities/wallet';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
